trustrail/backend/django_project/requirements.txt
asgiref==3.11.1
blinker==1.9.0
certifi==2026.4.22
charset-normalizer==3.4.7
click==8.1.8
contourpy==1.3.3
cs50==9.4.0
cycler==0.12.1
Django==6.0.2
django-cors-headers==4.9.0
djangorestframework==3.17.1
Flask==3.1.0
fonttools==4.61.1
idna==3.13
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.9
MarkupSafe==3.0.2
matplotlib==3.10.8
numpy==2.4.2
packaging==24.2
pillow==12.1.1
pyparsing==3.3.2
python-dateutil==2.9.0.post0
requests==2.33.1
six==1.17.0
SQLAlchemy==2.0.39
sqlparse==0.5.3
termcolor==2.5.0
tk==0.1.0
# Editable Git install with no remote (trustrail==0.1.0)
-e /Users/mac/Desktop/ideatho_n/trustrail
typing_extensions==4.12.2
urllib3==2.6.3
Werkzeug==3.1.3
wheel==0.45.1
whitenoise==6.12.0
